% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGseaNb.R
\name{new_gseaNb}
\alias{new_gseaNb}
\title{new_gseaNb}
\arguments{
\item{object}{GSEA enrichment results.}

\item{subPlot}{Integer indicating which plot to show: 1, 2, or 3. Default is 3.}

\item{lineSize}{Numeric indicating the size of the curve line. Default is 0.8.}

\item{lineAlpha}{Numeric indicating the alpha transparency for the curve line. Default is 1.}

\item{geneSetID}{Character vector specifying which pathway names to plot.}

\item{rmSegment}{Logical indicating whether to remove segments on the curve plot. Default is FALSE.}

\item{termWidth}{Integer specifying the width of the term name. Default is 40.}

\item{segCol}{Character specifying the segment color on the curves. Default is "red".}

\item{addGene}{Character vector of gene names to be added to the curve. Default is NULL.}

\item{geneCol}{Character specifying the color of gene name labels. Default is NULL.}

\item{arrowAngle}{Numeric indicating the angle of the arrow. Default is 20.}

\item{arrowLength}{Numeric indicating the length of the arrow line. Default is 0.2.}

\item{arrowEnd}{Character specifying the end of the arrow. Default is "first".}

\item{arrowType}{Character specifying the type of arrow. Default is "closed".}

\item{curveCol}{Character vector specifying the colors of the curves. Default is c("#76BA99", "#EB4747", "#996699").}

\item{htCol}{Character vector specifying the heatmap colors. Default is c("#148B77", "#F49E00").}

\item{rankCol}{Character vector specifying the gene rank fill colors. Default is c("#148B77", "white", "#F49E00").}

\item{rankSeq}{Integer specifying the breaks for the gene rank plot X-axis. Default is 5000.}

\item{htHeight}{Numeric indicating the relative height of the heatmap when `subPlot = 2`. Default is 0.3.}

\item{force}{Numeric indicating the gene label force, referring to the `geom_text_repel` function. Default is 20.}

\item{max.overlaps}{Integer indicating the maximum number of overlaps for gene labels, referring to the `geom_text_repel` function. Default is 50.}

\item{geneSize}{Numeric indicating the size of gene label text. Default is 4.}

\item{newGsea}{Logical indicating whether to show the new style of plot. Default is FALSE.}

\item{addPoint}{Logical indicating whether to add a point layer in the new style plot. Default is TRUE.}

\item{newCurveCol}{Character vector specifying the colors for the new style plot curve. Default is c("#336699", "white", "#993399").}

\item{newHtCol}{Character vector specifying the colors for the new style plot heatmap. Default is c("#336699", "white", "#993399").}

\item{rmHt}{Logical indicating whether to remove the heatmap in the new style plot. Default is FALSE.}

\item{addPval}{Logical indicating whether to add p-value and NES. Default is FALSE.}

\item{pvalX}{Numeric indicating the x-position for the p-value label. Default is 0.9.}

\item{pvalY}{Numeric indicating the y-position for the p-value label. Default is 0.9.}

\item{pvalSize}{Numeric indicating the size of the p-value label text. Default is 4.}

\item{pCol}{Character specifying the color of the p-value label. Default is "grey".}

\item{pHjust}{Numeric indicating the horizontal justification for the p-value label. Default is 1.}

\item{rmPrefix}{Logical indicating whether to remove the prefix like "GOBP/KEGG/CC/MF_*" from GO term names. Default is TRUE.}

\item{nesDigit}{Integer indicating the number of digits to retain for the NES score. Default is 2.}

\item{pDigit}{Integer indicating the number of digits to retain for the p-value and adjusted p-value. Default is 2.}

\item{markTopgene}{Logical indicating whether to add the top n genes on the plot. Default is FALSE.}

\item{topGeneN}{Integer indicating the number of top genes to be marked on the plot. Default is 5.}

\item{kegg}{Logical indicating whether the input is a `gseKEGG` object. Default is FALSE.}

\item{legend.position}{Character specifying the legend position. Default is "right".}

\item{add.geneExpHt}{Logical indicating whether to add a target gene expression heatmap. Default is FALSE.}

\item{exp}{Data frame containing the expression matrix in TPM/FPKM/RPKM format. Default is NULL.}

\item{scale.exp}{Logical indicating whether to scale the expression matrix. Default is TRUE.}

\item{sample.order}{Character vector specifying the sample order in the expression matrix. Default is NULL.}

\item{exp.col}{Character vector specifying the expression colors. Default is c("blue", "white", "red").}

\item{ht.legend}{Logical indicating whether to show the heatmap legend. Default is TRUE.}

\item{ght.relHight}{Numeric indicating the relative height of the gene expression heatmap to the main plot. Default is 0.4.}

\item{ght.geneText.size}{Numeric indicating the size of the gene label text in the heatmap. Default is 6.}

\item{ght.facet}{Logical indicating whether to facet the expression heatmap. Default is FALSE.}

\item{ght.facet.scale}{Character specifying the scale argument for the faceted plot. Default is "free".}

\item{termID.order}{Character vector specifying the facet term ID orders. Default is NULL.}

\item{rank.gene}{Character vector specifying the genes to label on the rank plot. Default is NULL.}

\item{rank.gene.nudgey}{Numeric indicating the y-axis nudge for gene labels on the rank plot. Default is 2.}

\item{plotHeightRatio}{Numeric vector specifying the height ratios of subplots. Default is c(0.5, 0.2, 0.3).}

\item{segmentSize}{Numeric indicating the size of the segments. Default is 0.8.}

\item{segmentAlpha}{Numeric indicating the alpha transparency for the segments. Default is 0.8.}
}
\value{
A ggplot2 object.
}
\description{
Generates GSEA (Gene Set Enrichment Analysis) plots with customizable parameters.
}
\examples{
\dontrun{
# Load data
test_data <- system.file("extdata", "gseaRes.RDS", package = "GseaVis")
gseaRes <- readRDS(test_data)

# All plots
new_gseaNb(object = gseaRes,
           geneSetID = 'GOBP_NUCLEOSIDE_DIPHOSPHATE_METABOLIC_PROCESS',
           subPlot = 2)
}
}
