% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCoreEnrichment.R
\name{process_core_enrichment}
\alias{process_core_enrichment}
\title{Process Core Enrichment Genes and Count Positive/Negative Correlations}
\usage{
process_core_enrichment(core_enrichment, correlation_data)
}
\arguments{
\item{core_enrichment}{A string of core enrichment genes separated by "/".}

\item{correlation_data}{A data frame containing gene correlations with columns "gene2" and "cor".}
}
\description{
This function processes core enrichment genes from a given string, identifies unique and non-unique genes,
and counts the number of positive and negative correlations for both unique and non-unique gene sets.
}
\examples{
\dontrun{
core_enrichment <- "gene1/gene2/gene3/gene1"
correlation_data <- data.frame(gene2 = c("gene1", "gene2", "gene3", "gene4"),
                               cor = c(0.5, -0.3, 0.7, 0.2))
process_core_enrichment(core_enrichment, correlation_data)
}
}
